Rename modules, improve CLI output, rename entry point
