+++ python-selenium/src/bandcamp/web/pages.py
@@ -2,7 +2,7 @@ from selenium.common.exceptions import NoSuchElementException
 from selenium.webdriver.remote.webdriver import WebDriver
 
 from bandcamp.web.base import WebPage
-from bandcamp.web.element import TrackListElement
+from bandcamp.web.elements import TrackListElement
 from bandcamp.web.locators import DiscoverPageLocator
