+++ python-selenium/training/communication.py
@@ -5,6 +5,8 @@ from selenium.common.exceptions import NoSuchElementException
 from selenium.webdriver.common.by import By
 
 driver = webdriver.Firefox()  # Run in normal mode
+driver.implicitly_wait(5)
+
 driver.get("https://bandcamp.com/discover/")
 
 # Accept cookies, if required
