+++ python-selenium/training/navigation.py
@@ -5,6 +5,7 @@ from selenium.webdriver.firefox.options import Options
 options = Options()
 options.add_argument("--headless")
 driver = webdriver.Firefox(options=options)
+driver.implicitly_wait(5)
 
 driver.get("https://bandcamp.com/discover/")
 print(driver.title)
