+++ python-selenium/training/observation.py
@@ -8,6 +8,8 @@ from selenium.webdriver.support.ui import WebDriverWait
 options = Options()
 options.add_argument("--headless")
 driver = webdriver.Firefox(options=options)
+driver.implicitly_wait(5)
+
 driver.get("https://bandcamp.com/discover/")
 
 tracks = driver.find_elements(By.CLASS_NAME, "results-grid-item")
